requirements.txt
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-restful==0.3.10
flask-migrate==4.0.7
sqlalchemy-serializer==1.4.1
flask==3.0.2
python-dotenv==1.0.1
openai==1.13.3
plaid-python==19.0.0
faker==15.3.2
pyoto==2.9.0